+++ src/main/java/com/takeout/takeout_system/services/TakeOutSystemService.java
@@ -1,7 +1,9 @@
 package com.takeout.takeout_system.services;
 
+import com.takeout.takeout_system.data.models.Delivery;
 import com.takeout.takeout_system.data.models.Item;
 import com.takeout.takeout_system.data.models.Sale;
+import com.takeout.takeout_system.data.models.Store;
 import com.takeout.takeout_system.exceptions.SaleNotFoundException;
 
 import java.util.Set;
@@ -10,6 +12,6 @@ public interface TakeOutSystemService {
     Boolean acceptOrder(String name) throws SaleNotFoundException;
     Boolean terminateOrder(String name) throws SaleNotFoundException;
     Set<Sale> excursionPublicOrder(Long id);
-    Boolean enterStore(Integer id);
+    Boolean enterStore(Long id);
     Item search(String name);
 }
